requirements_intel_windows.txt
anyio==4.11.0
certifi==2025.11.12
click==8.3.0
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
dpcpp-cpp-rt==2025.2.1
filelock==3.20.0
fsspec==2025.10.0
ftfy==6.3.1
h11==0.16.0
h3==4.3.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.4
idna==3.11
intel-cmplr-lib-rt==2025.2.1
intel-cmplr-lib-ur==2025.2.1
intel-cmplr-lic-rt==2025.2.1
intel-opencl-rt==2025.2.1
intel-openmp==2025.2.1
intel-pti==0.13.1
intel-sycl-rt==2025.2.1
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==2.1.5
mkl==2025.2.0
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
onemkl-sycl-blas==2025.2.0
onemkl-sycl-dft==2025.2.0
onemkl-sycl-lapack==2025.2.0
onemkl-sycl-rng==2025.2.0
onemkl-sycl-sparse==2025.2.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
python-dateutil==2.9.0.post0
pytorch-triton-xpu==3.5.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
scikit-learn==1.7.2
scipy==1.16.3
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tbb==2022.2.0
tcmlib==1.4.0
threadpoolctl==3.6.0
torch==2.9.1+xpu
torchaudio==2.9.1+xpu
torchvision==0.24.1+xpu
tqdm==4.67.1
typer-slim==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
umf==0.11.0
wcwidth==0.2.14
